Retreive all matching packages by paging
A JSON array should retain the same order. The TypeSet was not working
correctly, only resulting in the first list entry in the data-source. It
seems that I should have used the TypeList anyway as the array should be
ordered.

The Cloudsmith API only returns a maximum of 100 results per page. This
means to obtain all the packages, I need to retrieve the other pages.
This is fairly straightforward using the header fields to determine the
number of pages.