Move terraform.ResourceProvider usage up to `main.go`
This is a small refactoring, done in service of porting this provider
to [Pulumi][pulumi] using their [Pulumi/Terraform
Bridge][tfbridge]. The bridge introspects the Terraform provider's Go
code to extract the schema of the provider as an input to a codegen
process. It is coded in terms of `schema.Provider`, rather than
`terraform.ResourceProvider`, though.

[pulumi]: https://pulumi.com
[tfbridge]: https://github.com/pulumi/pulumi-terraform-bridge